move collective fusion to separate file
